refactored repo and adding packaging configuration
