+++ setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = autoid3
+version = 1.0.0
+author = Erik Tyryshkin
+description = Automatically add title, artist, album, genre, and cover art metadata to supported mp3 tracks.
+long_description = file: README.md
+url = https://github.com/kingcobra2468/AutoID3
+classifiers =
+    Programming Language :: Python :: 3
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.7
+install_requires =
+    shazamio==0.1.1.0
+    eyed3==0.9.6
+    aiohttp==3.8.1
+    pydub==0.24.1
+    httpx==0.22.0
+
+[options.packages.find]
+where = src
